rename: Diary 조회 클래스명 변경
