+++ src/main/java/flab/nutridiary/diary/controller/DiaryController.java
@@ -7,7 +7,7 @@ import flab.nutridiary.diary.dto.response.DiarySavedResponse;
 import flab.nutridiary.diary.dto.response.query.DiaryRecordWithProduct;
 import flab.nutridiary.diary.service.AddDiaryRecordService;
 import flab.nutridiary.diary.service.DiaryRegisterService;
-import flab.nutridiary.diary.service.DiaryRetrievalService;
+import flab.nutridiary.diary.service.DiarySearchService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -20,7 +20,7 @@ import java.util.List;
 public class DiaryController {
     private final DiaryRegisterService diaryRegisterService;
     private final AddDiaryRecordService addDiaryRecordService;
-    private final DiaryRetrievalService diaryRetrievalService;
+    private final DiarySearchService diarySearchService;
 
     @PostMapping("/diary/new")
     public ApiResponse<DiarySavedResponse> createDiary(@Valid @RequestBody DiaryRegisterRequest diaryRegisterRequest) {
@@ -36,6 +36,6 @@ public class DiaryController {
     @GetMapping("/diary")
     public ApiResponse<List<DiaryRecordWithProduct>> getDiary(@RequestParam(name = "diaryDate") LocalDate diaryDate) {
         Long memberId = 1L;
-        return ApiResponse.success(diaryRetrievalService.getDiary(memberId, diaryDate));
+        return ApiResponse.success(diarySearchService.getDiary(memberId, diaryDate));
     }
 }
