+++ src/main/java/flab/nutridiary/diary/repository/DiarySearchRepository.java
@@ -5,6 +5,6 @@ import flab.nutridiary.diary.dto.response.query.DiaryRecordWithProduct;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface DiaryRetrievalRepository {
+public interface DiarySearchRepository {
     List<DiaryRecordWithProduct> findDiaryWithProductAllByMemberIdAndDiaryDate(Long memberId, LocalDate diaryDate);
 }
